Update API: ChangePassword, Logout; Add 'Mobile Number' column to Contact Us; Modify Workshop Orders Method

+++ src/modules/user/user.service.ts
@@ -710,12 +710,21 @@ export class UserService {
   }
 
   async logout(user_id: number): Promise<Response> {
+    const user = await this.deviceUserRepository.findOne({
+      where: { user_id: user_id },
+    });
+
+    if (!user) {
+      throw new NotFoundException('user not found');
+    }
+
     await this.deviceUserRepository.delete({
       user_id: user_id,
     });
     return {
       statusCode: 200,
       message: 'logout successfully',
+      data: user,
     };
   }
